updated Xlw plus stuff to allow easy addition of new types also made work with 7.1 as well as 8 and devcpp (6 to do)
git-svn-id: https://svn.code.sf.net/p/xlw/code/trunk/xlw@283 948fe115-5b3c-0410-8233-88faca9dced8

+++ xlw/XlfExcel.inl
@@ -86,7 +86,7 @@ INLINE LPSTR XlfExcel::GetMemory(size_t bytes)
     XlfBuffer& buffer = freeList_.front();
     if (offset_ + bytes < buffer.size)
     {
-      int temp = offset_;
+      int temp = static_cast<int>(offset_);
       offset_ += bytes;
       return buffer.start + temp;
     }
